+++ CHANGELOG.md
@@ -3,6 +3,9 @@
 ## 2.3.5 Timeless
 
 - time breakpoints since document load
+- day time classes (day-time-5h, day-time-17h, ...)
+- day period (day-period-morning, day-period-evening, ...)
+- year period (year-period-summer, year-period-winter, ...)
 
 ## 2.3.4 Newborn